pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "warning-agent"
version = "0.1.0"
description = "Narrow alert intelligence agent for Prometheus + SigNoz evidence, packetization, triage, investigation, and markdown reporting."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "Peng" }
]
dependencies = [
  "fastapi>=0.116,<1",
  "uvicorn>=0.35,<1",
  "pydantic>=2.11,<3",
  "httpx>=0.28,<1",
  "jinja2>=3.1,<4",
  "pyyaml>=6.0,<7",
  "scikit-learn>=1.7,<2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9",
  "ruff>=0.12,<1",
  "jsonschema>=4.24,<5",
]

[project.scripts]
warning-agent = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-ra"
